backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
transformers==4.35.0
torch==2.0.1
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
safetensors==0.3.2
sentencepiece==0.2.0 